fix consquetive calculations
Debug time info:
Can not calculate three consquetive calculations
(without using = between them).
it seems that we lose the first binary operation's result
when clicking the second binary operator.
after one calculation and second input of an operator and
at the third input of a number the getDigit function
clears the screen, so the previous result is lost.